+++ R/availableCores.R
@@ -100,11 +100,12 @@
 #'  \item `"_R_CHECK_LIMIT_CORES_"` -
 #'    Query environment variable \env{_R_CHECK_LIMIT_CORES_} (logical or
 #'    `"warn"`) used by `R CMD check` and set to true by
-#'    `R CMD check --as-cran`. If set to a non-false value, then a maximum
+#'    `R CMD check --as-cran`. In addition, package \pkg{parallelly} sets
+#'    `_R_CHECK_LIMIT_CORES_=true` when _loaded_ if it detects that `R CMD`
+#'    is running and _builds_ package vignettes via `R CMD build` and
+#'    `R CMD check`, which is something `R CMD check` does not do itself.
+#'    If `_R_CHECK_LIMIT_CORES_` is set to a non-false value, then a maximum
 #'    of 2 cores is considered.
-#'    Note that `_R_CHECK_LIMIT_CORES_` is _not_ set when `R CMD build`
-#'    builds vignettes or when `R CMD check --as-cran` re-builds then as
-#'    part of the package checks.
 #'
 #'  \item `"Bioconductor"` -
 #'    Query environment variable \env{IS_BIOC_BUILD_MACHINE} (logical)