+++ R/zzz.R
@@ -6,6 +6,9 @@
   if (debug) options(parallelly.debug = TRUE)
   debug <- getOption("parallelly.debug", debug)
 
+  ## Drop CLI arguments
+  if (interactive()) cli_prune()
+
   ## Automatically play nice when 'R CMD check' runs?
   if (isTRUE(as.logical(getEnvVar2("R_PARALLELLY_R_CMD_CHECK_NICE", "TRUE"))) &&
       (inRCmdCheck() || inRCmdVignette())) {